+++ src/libs/amqp.server.ts
@@ -24,7 +24,11 @@ export const initAMQP = () => {
 
         const exchangeName = 'evolution_exchange';
 
-        channel.assertExchange(exchangeName, 'topic', { durable: false });
+        channel.assertExchange(exchangeName, 'topic', {
+          durable: true,
+          autoDelete: false,
+        });
+
         amqpChannel = channel;
 
         logger.info('AMQP initialized');